kalkulator 1

+++ src/math.kt
@@ -1,9 +1,9 @@
 fun main(args:Array<String>){
     //simple calc
     println("Vnesi go prviot broj:  ")
-    var n1:int= readLine()!!.toInt()
+    var n1:Int= readLine()!!.toInt()
     println("vnesi go vtoriot borj: ")
     var n2:Int= readLine()!!.toInt()
-    var sum:Float=n1+n2
+    var sum:Float=n1.toFloat() + n2.toFloat()
     println(sum)
 }
